+++ Akane/Akane/Binding/Observation/ViewModelObservation.swift
@@ -14,30 +14,42 @@ public class ViewModelObservation<ViewModelType: ComponentViewModel> : Observati
     public var next: [((ViewModelType) -> Void)] = []
 
     unowned let container: ComponentContainer
+    let observer: ViewObserver
 
-    public init(value: ViewModelType?, container: ComponentContainer) {
+    public init(value: ViewModelType?, container: ComponentContainer, observer: ViewObserver) {
         self.container = container
         self.value = value
+        self.observer = observer
     }
 }
 
 extension ViewModelObservation {
 
-    public func bind<ViewType: UIView>(to view: ViewType?) where ViewType: ComponentDisplayable {
+    public func bind<ViewType: Hashable & ComponentDisplayable>(to view: ViewType?) where ViewType.Parameters == ViewModelType {
         if let view = view {
             self.bind(to: view)
         }
     }
 
-    public func bind<ViewType: UIView>(to view: ViewType) where ViewType: ComponentDisplayable {
-        let controller: AnyComponentController? = self.container.component(for: view, createIfNeeded: true)
-
-        guard (controller != nil) else {
-            return
-        }
+    public func bind<ViewType: Hashable & ComponentDisplayable & Wrapped>(to view: ViewType)
+        where ViewType.Parameters == ViewModelType, ViewType.Wrapper.Parameters == ViewModelType {
+        let observer = self.observer.observer(identifier: view.hashValue)
+        let controller = self.container.component(for: view)
 
         self.observe { value in
-            controller!.setup(viewModel: value)
+            observer.dispose()
+            value.mountIfNeeded()
+            controller.bindings(observer, params: value)
+        }
+    }
+
+    public func bind<ViewType: Hashable & ComponentDisplayable>(to view: ViewType) where ViewType.Parameters == ViewModelType  {
+        let observer = self.observer.observer(identifier: view.hashValue)
+
+        self.observe { value in
+            observer.dispose()
+            value.mountIfNeeded()
+            view.bindings(observer, params: value)
         }
     }
 }