+++ Akane/Akane/Component/View/ContentDisplayable.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+public protocol ContentReferencer {
+    associatedtype Content
+
+    var content: Content { get }
+
+    init(content: Content)
+}
+