+++ Example/Example/Authors/AuthorsView.swift
@@ -24,8 +24,8 @@ class AuthorsView : UITableView, Displayable {
         self.dataSource = self
     }
     
-    func bindings(_ observer: ViewObserver, props: [Author]) {
-        self.authors = props
+    func bindings(_ observer: ViewObserver, params: [Author]) {
+        self.authors = params
         self.observer = observer
 
         self.reloadData()
@@ -43,7 +43,7 @@ extension AuthorsView : UITableViewDelegate, UITableViewDataSource {
 
         self.observer
             .observe(author)
-            .display(in: cell)
+            .bind(to: cell)
 
         return cell
     }